+++ server/core_notification.go
@@ -20,7 +20,10 @@ import (
 	"database/sql"
 	"encoding/base64"
 	"encoding/gob"
+	"encoding/json"
+	"errors"
 	"fmt"
+	"github.com/heroiclabs/nakama-common/runtime"
 	"time"
 
 	"github.com/gofrs/uuid/v5"
@@ -318,6 +321,89 @@ SELECT
 	if _, err := db.ExecContext(ctx, query, ids, userIds, subjects, contents, codes, senderIds); err != nil {
 		logger.Error("Could not save notifications.", zap.Error(err))
 		return err
+	}
+
+	return nil
+}
+
+func NotificationsGetId(ctx context.Context, logger *zap.Logger, db *sql.DB, userID string, ids ...string) ([]*runtime.Notification, error) {
+	if len(ids) == 0 {
+		return []*runtime.Notification{}, nil
+	}
+
+	for _, id := range ids {
+		if _, err := uuid.FromString(id); err != nil {
+			return nil, errors.New("expects id to be a valid id string")
+		}
+	}
+
+	params := []any{ids}
+	query := "SELECT id, user_id, subject, content, code, sender_id, create_time FROM notification WHERE id = any($1)"
+	if userID != "" {
+		query += " AND user_id = $2"
+		params = append(params, userID)
+	}
+
+	rows, err := db.QueryContext(ctx, query, params...)
+	if err != nil {
+		logger.Error("failed to list notifications by id", zap.Error(err))
+		return nil, fmt.Errorf("failed to list notifications by id: %s", err.Error())
+	}
+
+	defer rows.Close()
+
+	notifications := make([]*runtime.Notification, 0, len(ids))
+	for rows.Next() {
+		no := &runtime.Notification{Persistent: true, CreateTime: &timestamppb.Timestamp{}}
+		var createTime pgtype.Timestamptz
+		var content string
+		if err := rows.Scan(&no.Id, &no.UserID, &no.Subject, &content, &no.Code, &no.Sender, &createTime); err != nil {
+			_ = rows.Close()
+			logger.Error("Failed to scan notification from database.", zap.Error(err))
+			return nil, err
+		}
+		no.CreateTime.Seconds = createTime.Time.Unix()
+
+		var contentMap map[string]any
+		if err = json.Unmarshal([]byte(content), &contentMap); err != nil {
+			logger.Error("Failed to unmarshal notification content", zap.Error(err))
+			return nil, err
+		}
+		no.Content = contentMap
+
+		if no.Sender == uuid.Nil.String() {
+			no.Sender = ""
+		}
+		notifications = append(notifications, no)
+	}
+
+	return notifications, nil
+}
+
+func NotificationsDeleteId(ctx context.Context, logger *zap.Logger, db *sql.DB, userID string, ids ...string) error {
+	if len(ids) == 0 {
+		// NOOP
+		return nil
+	}
+
+	for _, id := range ids {
+		if _, err := uuid.FromString(id); err != nil {
+			return errors.New("expects id to be a valid uuid")
+		}
+	}
+
+	if userID != "" {
+		uid, err := uuid.FromString(userID)
+		if err != nil {
+			return errors.New("expects id to be a valid uuid")
+		}
+
+		return NotificationDelete(ctx, logger, db, uid, ids)
+	}
+
+	if _, err := db.QueryContext(ctx, "DELETE FROM notification WHERE id = any($1)", ids); err != nil {
+		logger.Error("failed to delete notifications by id", zap.Error(err))
+		return fmt.Errorf("failed to delete notifications: %s", err.Error())
 	}
 
 	return nil