+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project are documented below.
 The format is based on [keep a changelog](http://keepachangelog.com) and this project uses [semantic versioning](http://semver.org).
 
 ## [Unreleased]
+### Added
+- New runtime functions to get and delete notifications by id.
+
 ### Fixed
 - Correctly wire Go runtime shutdown function context.
 - Fix friends of friends api error when user has no friends.